\name{memm}
\alias{memm}
\title{Estimation framework for high-dimensional multivariate mediation analysis, integrating exposure, mediator, and outcome data through a penalized joint model solved by ADMM (Alternating Direction Method of Multipliers)}
\description{
}
\usage{

}
\arguments{
 }
\details{
}
\value{
}
\examples{
}
\seealso{

}
\author{
Neng Wang
}
